Fix voting concurrency with high volume polls
Polls that got high traffic in a short period of time "lost" votes. The
votes went through, but the counts didn't always update. We were pulling
the poll data in the `.before`, but simply doing a `votes++` would mean
it could overwrite other votes if someone else voted at the same time.
Switched to using atomic updates for Mongo which fixes this.